Moved constant and helpers one level up
