+++ common/v2/features/DownloadApp/DownloadApp.tsx
@@ -6,8 +6,9 @@ import cloneDeep from 'lodash/cloneDeep';
 import { ContentPanel } from 'v2/components';
 import { GithubService, AnalyticsService, ANALYTICS_CATEGORIES } from 'v2/services';
 import { OS } from 'v2/services/Github';
-import { DOWNLOAD_PAGE_URL, GITHUB_RELEASE_NOTES_URL } from './constants';
-import { getFeaturedOS } from './helpers';
+import { DOWNLOAD_PAGE_URL } from './constants';
+import { GITHUB_RELEASE_NOTES_URL } from 'v2/features/constants';
+import { getFeaturedOS } from 'v2/features/helpers';
 import { Layout } from 'v2/features';
 import { AppDownloadItem } from './types';
 import translate from 'translations';
