+++ common/v2/features/DownloadApp/constants.ts
@@ -1,2 +1,1 @@
 export const DOWNLOAD_PAGE_URL = 'https://download.mycrypto.com ';
-export const GITHUB_RELEASE_NOTES_URL = 'https://github.com/MyCryptoHQ/MyCrypto/releases/latest';
