+++ common/v2/features/constants.ts
@@ -1,0 +1,1 @@
+export const GITHUB_RELEASE_NOTES_URL = 'https://github.com/MyCryptoHQ/MyCrypto/releases/latest';
